train/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=8.0.0
numpy>=1.21.0
PyYAML>=5.4.0

# 高性能图像处理
opencv-python>=4.5.0

# 高效数据增强
albumentations>=1.3.0

# 现代backbone模型
timm>=0.9.0

# 可选：高级优化器
# lion-pytorch>=0.1.2  # Lion优化器
# adan-pytorch>=1.0.0  # Adan优化器

# 可选：Flash Attention（需要CUDA）
# flash-attn>=2.0.0
